ml_scanner_server/requirements.txt
attrs==24.3.0
bidict==0.23.1
blinker==1.9.0
cattrs==24.1.2
certifi==2024.12.14
charset-normalizer==2.0.12
click==8.1.8
coremltools==8.1
exceptiongroup==1.2.2
Flask==2.0.1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
numpy==2.0.1
packaging==24.2
Pillow==8.3.1
protobuf==3.19.0
pyaml==25.1.0
pyserial==3.5
python-engineio==4.11.2
python-socketio==5.12.1
PyYAML==6.0.2
requests==2.26.0
simple-websocket==1.1.0
sympy==1.13.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==2.0.1
wsproto==1.2.0
zipp==3.21.0
